+++ src/common/state/GaussianControlPointList.C
@@ -672,7 +672,7 @@ GaussianControlPointList::RemoveControlPoints(int index)
 int
 GaussianControlPointList::GetNumControlPoints() const
 {
-    return controlPoints.size();
+    return (int)controlPoints.size();
 }
 
 // ****************************************************************************
